Add support for Rails 7+
The `searchable` module broke on Rails 7 because it was explicitly looking for Rails 5 or 6. The behavior was actually the same between 5 and 6, and works the same way on Rails 7, so I've simplified the logic for which class methods to include: all of them. Rails versions older than 5.1 are already incompatible with the gemspec, so this change should keep the gem working through major version changes in the future.